add more debug info

+++ mmdet/models/roi_heads/bbox_heads/iclip_bbox_head.py
@@ -156,7 +156,7 @@ class IclipBBoxHead(BBoxHead):
                     losses.update(acc_)
                 else:
                     # losses['acc'] = accuracy(cls_score, labels)
-                    losses['acc'] = (argmax == labels).sum()
+                    losses['acc'] = (argmax == labels).sum().to(torch.float32)
         if bbox_pred is not None:
             bg_class_ind = self.num_classes
             # 0~self.num_classes-1 are FG, self.num_classes is BG
